Handle missing network file

+++ app/controllers/network_controller.rb
@@ -11,34 +11,39 @@ class NetworkController < ApplicationController
   def edit
     tmp = Tempfile.new("temp_vars")
 
-    file_data.each do |line|
-      if line.include?("INTERNAL") || line.include?("EXTERNAL")
-        content = line.split("=")
-        variable = content[0].remove("export ")
-        original_value = content[1].split('"')[1]
+    if File.exist? network_variables
+      file_data.each do |line|
+        if line.include?("INTERNAL") || line.include?("EXTERNAL")
+          content = line.split("=")
+          variable = content[0].remove("export ")
+          original_value = content[1].split('"')[1]
 
-        if original_value.empty?
-          line = line.insert(line.index('"') + 1, network_params[variable])
-        else
-          line = line.sub original_value, network_params[variable]
+          if original_value.empty?
+            line = line.insert(line.index('"') + 1, network_params[variable])
+          else
+            line = line.sub original_value, network_params[variable]
+          end
         end
+
+        tmp << line
       end
 
-      tmp << line
-    end
+      tmp.close
 
-    tmp.close
-    if run_shell_command("cp --no-preserve=mode,ownership #{tmp.path} #{network_variables}")
-      if run_global_script(ENV['NETWORK_SET'])[:status].success?
-        flash[:success] = 'Network configuration successfully modified'
+      if run_shell_command("cp --no-preserve=mode,ownership #{tmp.path} #{network_variables}")
+        if run_global_script(ENV['NETWORK_SET'])[:status].success?
+          flash[:success] = 'Network configuration successfully modified'
+        else
+          flash[:danger] = 'Encountered an error whilst trying to run the setup script'
+        end
       else
-        flash[:danger] = 'Encountered an error whilst trying to run the setup script'
+        flash[:danger] = 'Encountered an error whilst trying to modify the network configuration'
       end
-    else
-      flash[:danger] = 'Encountered an error whilst trying to modify the network configuration'
-    end
 
-    tmp.delete
+      tmp.delete
+    else
+      flash[:danger] = 'No network variables file defined'
+    end
 
     redirect_to network_path
   end
